viceroyfs: simplify how extra mounts are made on top of chroot
This commit uses a child process to create the necessary mounts on top
of the chroot environment.

We change the main handler from a lazy handler to the voverlay. A lazy
handler will reset the ID cache every time a new filesystem is
introduced and cause collisions at the FS level. Using a lazy handler
only for the underlying filesystem of the overlay should be handled
more gracefully.